Client part of JOO-49 Allow EXML to append / prepend to Array-valued config properties http://jangaroo.myjetbrains.com/youtrack/issue/JOO-49
New helper function Exml.apply() that replaces Ext.apply() and
can handle append / prepend of Array config values as specified
by an additional property <name>$at which determines the
index at which the existing list is to be inserted into the given list,
where -1 means "length of the given list".
There are constants APPEND (0) and PREPEND (-1) for the
corresponding indexes to use for appending and prepending
the given list to the existing list.