fixes LIBS-45 Computation of DisplayObjectContainer bounds incorrect http://jangaroo.myjetbrains.com/youtrack/issue/LIBS-45
Evaluation: actually two bugs.
First one is that if no child with width or height > 0 exists,
left / right / top / bottom are never assigned and the result
is wrong.
Second one is that Number.MIN_VALUE in ActionScript is
not the smallest number that can be represented as a Number,
but the smallest number *greater than zero* that can be
represented as a Number! So we must use -Number.MAX_VALUE
instead of Number.MIN_VALUE.