#LIBS-48 Fixed stage.width / stage.height temporarily zero when using width/height 100% and cacheAsBitmap http://jangaroo.myjetbrains.com/youtrack/issue/LIBS-48
When stage.cacheAsBitmap = true is set in the main class of the
application, the stage has already been created. Thus, it is reset
and recreated as a <canvas> element. JooFlash tried to measure
the dimensions of the new element before adding it to the DOM.
Now, the element replaces the old stage element first, then, its
dimensions are measured.