Browser JavaScript "interfaces" also have to be declared [Native]. Found some inconsistent names and renamed to actual browser names: * Collection => NodeList * Style => CSS2Properties (Firefox) / MSStyleCSSProperties (IE) * ControlRange => ControlRangeCollection (IE only; no longer inherits from TextRange) * IESelection => MSSelection (IE only)
To allow CSS2Properties to have a different name in IE, introduced a new
syntax for "native!" plugin: you can separate fall-back alternatives with "|".

Note that most interfaces are not supported in IE8, so if you want to
build IE8-compatible applications, do not use them in "as" or "is" expressions
(simple type casts are currently okay).