fixes JOO-84 Using field initializer may lead to NPE in IE11 http://jangaroo.myjetbrains.com/youtrack/issue/JOO-84
One more indirection in calling the initializer function keeps
IE11 from mis-guessing the type and running into the error
described in the ticket.
Also, removed the old workaround for Firefox 9. The new
workaround should fix FF9, too, and even if it doesn't, FF9
is sufficiently old to no longer be supported by Jangaroo.