Adapt Ext config classes to latest MXML pattern.
To use non-config classes in MXML, they now have to be contained in manifest.xml
and provide a hidden "config" write-only property of their config class type, which is
also declared as [DefaultProperty("config")].
This is used as a replacement for calling the constructor with a config parameter.

Partially reverts 1168d05f  More general type for configs that allow config and target objects
for certain properties.
This is no longer needed, as you can now use target classes as "type casts", for example
to fill a <baseAction> property.