Support Flex events in MXML and ActionScript.
The compiler now produces working code when using Flex-style events
in MXML or ActionScript. The limitation for ActionScript is that the
new method IObservable#addEventListener() must be used with an
event name specified as a constant from the corresponding event class.

The old joo.addEventListener() has been replaced by AS3.addEventListener()
with a slightly different signature that more closely resembles the signature
of IObservable#addEventListener().