Removing "AS3." namespace prefix.
Matching the change in jangaroo-tools, the "AS3." prefix
has to be removed in all code that refers to JavaScript runtime
names.
The [Native] annotation of getTimer.as now work without any
name, because the default is that the ActionScript and JavaScript
name are the same, which now again holds true.
Everything looks much more elegant now! :-)