CKEditor 4.5.7 to 4.7.1: Additions
This commit is based on a copy of the sources of the CKEditor 4.7.1
release available at https://github.com/ckeditor/ckeditor-dev.

It is the first of three commits only containing the unmodified files
from CKEditor without any patches we previously applied to 4.5.7 like
CSP fixes.

This one, the third commit, contains all files which were added to
4.7.1 compared to our (patched) 4.5.7 release. Thus it especially also
contains any plugins we previously removed.