CKEditor 4.5.7 to 4.7.2: Fake CKEditor Version
In order to fake a non-dev CKEditor version you currently have to
add a git tag. The build_ckeditor_from_source.sh script now creates
such a fake take prior to triggering the build process and removes
the tag eventually.

The version tag contains a jangaroo-postfix. This is ignored by
current CKEditor builder, so the resulting version is "4.7.2". We
still keep this postfix as at the end it would be nice if CKEditor
version contains a reference to Jangaroo.