Fixes LIBS-83: Update Jangaroo Browser API
https://jangaroo.myjetbrains.com/youtrack/issue/LIBS-83

For details, see JooTrack issue.
The only technical detail not described there is that the type of
Window#frames is a made-up type "FrameList" that is actually, like
in all browser implementations I checked, the window itself! 8-O
This can be expressed in Jangaroo by [Native("Window")], so that
the ActionScript expression "windows.frames is FrameList" will
actually work.