Rename ... 'rest' parameter to 'arguments'
Then, Jangaroo magic is applied and the JavaScript 'arguments'
value is used directly (as long as no Array method is called
on the parameter).
When converting to TypeScript, to satisfy strict mode, 'arguments'
is renamed to 'arguments$', but Babel renames it back to
'arguments', resulting in the exact same JS code.