Implement LIBS-88: Lazy Initialization for Singletons
https://jangaroo.myjetbrains.com/youtrack/issue/LIBS-88

This is the runtime counterpart of JOO-152. It implements
a new Ext class extension __lazyFactory__ that is similar
to __factory__, but invokes the initializer function on
first access (lazy).

Like for static properties, used accessors that call
the factory function on first access.
Note that if the singleton is set from outside before
having been read at all, I chose not to call the
initializer, because its result would be overwritten
afterwards, anyway. This might come unexpected if the
initializer code has side-effects.
In debug mode (#joo.debug), all lazy singleton
initializations are logged to the console.