Workaround for @see reference to package-scope function
IDEA wants the simple syntax "js.fetch()", but that does not
work in "real" ASDoc.
ASDoc wants "js.#fetch()" to render the correct link, but then
renders the link test incorrectly as "js..fetch()", so we repeat
"js.fetch()" as link text.
Now we end up with one link "js.fetch() (workaroun for IDEA)"
that only works in IDEA and renders as text in ASDoc HTML, and
one link "js.fetch()" that renders and works correctly in ASDoc.