trace() fix: detect log level also in multi-line messages
Unfortunately, '.' in a JavaScript RegExp does not match new-line
characters ('\n'), so the pattern to detect whether the message
string starts with a log level does not match.
There is an option 's' that can be appended to the RegExp to
let '.' match really any character, but that is not supported
in IE11.
What _does_ work is the character class '\s\S', which means
"any whitespace character OR any non-whitespace character".