[jangaroo-browser] Clean up Event API
There were several outdated and incorrectly assigned methods
in the jangaroo-browser Event API. Some may have been that
way because of browser incompatibilities, but the oldest
browser we need to support now is IE11, so if even IE11
implements the standard, the AS3 API has been updated
accordingly.

* Most methods assigned to js.Event actually belong to
  a subclass and have been moved accordingly
* The key event class is actually called KeyboardEvent
  and does not contain any of the DOM_VK_... constants.
  This is an outdated Firefox specialty, but is kept to
  maintain AS3 API compatibility. KeyEvent now extends
  KeyboardEvent, so that it still works as the type of
  events received in keyboard event listeners.
* Although MouseEvent and KeyboardEvent share some
  properties (shiftKey, ctrlKey, altKey, metaKey,
  getModifierState()), in the standard and in all browser
  implementations, these are not inherited, but defined
  twice. shiftLeft and ctrlLeft are obsolete and have
  been removed. If you want to find out exactly which
  shift / ctrl key has been pressed, use
  KeyboardEvent#location with the constants defined in
  that class.
* MouseEvent and KeyboardEvent now inherit from UIEvent,
  not only from Event.
* There are a few more UIEvent subclasses that are not
  yet represented in AS3: FocusEvent, WheelEvent,
  InputEvent. If you need these, you can use UIEvent
  and access additional properties in an untyped fashion.

The only usage of these Event classes that had to be
adapted was in flash.display.Stage, which is the adapter
from browser events to emulated Flash events. Although
this library is not maintained anymore, these changes
were necessary to make it compile.