make sure that code in jangaroo-runtime and jangaroo-browser is loaded earlier
due to the introduction of [Lazy] annotation it can happen that code that requires stuff from *.module.js ist loaded before the module.
Made sure that loading the polyfills / API will be performed as soon as anything in jangaroo-runtime is loaded.

On the fly made sure that the polyfills use the (private) loadScriptsSync instead of loadScriptSync, otherwise the poly fills are being loaded too late.