[jangaroo-runtime] Change type of Object#constructor from Object to Function
This is the reality in JavaScript. Only in "real" ActionScript, a class
is not represented by its constructor function, but by some class object.
Note that while this fixes the API, this is a breaking change. Since Function,
unlike Object, is not a 'dynamic' class in ActionScript, untyped access of
constructor any properties is no longer allowed. You now must use indexed
access for that. For example,

   this.constructor.someStaticMember

has always been an untyped access to 'someStaticMember', but worked at
run-time (if you did not misspell the member). Now, emphasizing the
untyped access, you must write

  this.constructor["someStaticMember"]