[runtime,browser] Fix: Adding [Native("...")] annotations for keyword-named methods makes joo.aliasKeywordMembers obsolete
For some native JS methods named 'delete' which had to be aliased to
'delete_' in ActionScript, the [Native("delete")] annotation was missing.
With this annotation present, it is no longer necessary to patch the
corresponding JavaScript classes with '_' post-fixed members. This was
done by the utility function joo.aliasKeywordMembers, which is now gone.

This also fixes usages of these '_' post-fixed methods when converting
to TypeScript.

Note that this change _may_ break your code, if you used untyped access
to an '_' post-fixed method (which does not really make sense, as you
could have used foo["delete"]() or foo["catch"]() in the first place),
or used the '_' post-fixed methods from JavaScript (same).