[jangaroo-browser] Add constructors for Event and its subclasses
The default in ActionScript is a no-arg constructor, but JS Event
even requires its first constructor argument and has an optional
second argument. This holds true for all of its subclasses.
While untyped access allows to call any missing methods, it is
more complicated to do an untyped constructor call, so fixing the
constructor signatures is important.
As a workaround, you can use
    window.document.createEvent(eventType)
but using the Event constructors is the recommended "modern" way
(see MDN).