CKEditor: Remove Patch from WYSIWYG-Area
The patch was only required for running CKEditor in IFrame and providing
a CSP Security Header Token.

As the default deployment in CoreMedia Studio does not use CKEditor
embedded as IFrame anymore, there is no need to keep up this patch.
Especially, because it clashes with recent change to CKEditor
4.17.1.