[JooFlash] Move jooflash-core back into jooflash
To make jooflash more independent of jangaroo-libs,
re-integrating jooflash-core into jooflash.
The separation was made to get a smaller footprint
for FlexUnit, which only uses a very small subset
of the Flash API. For FlexUnit/JooUnit, we are going
to maintain jooflash-core, but there is no need to
share the code with jooflash. On the contrary, this
would only mean there is a risk to break JooUnit
when changing code in jooflash-core for jooflash.